backend/app/db/__init__.py
"""
Database Module - Firestore Connection (Future Use)
Reserved for user data, history, and analytics storage
"""

from google.cloud import firestore
import os

# Initialize Firestore client (when needed)
def get_firestore_client():
    """
    Get Firestore database client
    Requires GOOGLE_APPLICATION_CREDENTIALS environment variable
    """
    try:
        return firestore.Client()
    except Exception as e:
        print(f"Warning: Firestore not configured: {e}")
        return None

db = None

def init_db():
    """Initialize database connection"""
    global db
    db = get_firestore_client()
    return db


